+++ .travis.yml
@@ -7,6 +7,8 @@ sudo: false
 cache:
   directories:
     - node_modules
+before_script:
+- npm run bootstrap
 addons:
   apt:
     packages:
