+++ package.json
@@ -41,6 +41,8 @@
     "jsdoc": "^3.5.5",
     "lerna": "^2.8.0",
     "rollup": "^0.68.2",
+    "rollup-plugin-json": "^3.1.0",
+    "rollup-plugin-node-resolve": "^4.0.0",
     "semistandard": "*",
     "snazzy": "^8.0.0",
     "tape": "^4.0.0",
