+++ packages/common/package.json
@@ -1,21 +1,15 @@
 {
-  "name": "@terraformer/arcgis",
-  "description": "Tools to convert ArcGIS JSON geometries to GeoJSON geometries and vica-versa.",
+  "name": "@terraformer/common",
+  "description": "Utility methods used in more than one @terraformer package.",
   "version": "2.0.0",
-  "author": "Patrick Arlt <patrick.arlt@gmail.com>",
+  "author": "John Gravois <john@esri.com>",
   "bugs": {
     "url": "https://github.com/terraformer-js/terraformer/issues"
   },
   "contributors": [
     "John Gravois <john@esri.com>"
   ],
-  "dependencies": {
-    "@terraformer/common": "2.0.0"
-  },
-  "files": [
-    "index.js",
-    "dist/arcgis.umd.js.map"
-  ],
+  "files": [ "index.js" ],
   "homepage": "https://github.com/terraformer-js/terraformer",
   "keywords": [
     "arcgis",
@@ -25,8 +19,7 @@
     "geometry"
   ],
   "license": "MIT",
-  "main": "dist/arcgis.umd.js",
-  "unpkg": "dist/arcgis.umd.js",
+  "main": "index.js",
   "module": "index.js",
   "jspm": {
     "main": "index.js",
@@ -37,8 +30,8 @@
     "type": "git",
     "url": "https://github.com/terraformer-js/terraformer"
   },
-  "scripts": {
-    "build": "rollup -c ../../rollup.config.js",
-    "prepare": "npm run build"
+  "scripts": {},
+  "browserify": {
+    "transform": [["babelify", { "presets": ["@babel/preset-env"] }]]
   }
 }
